Utils.js to handle generic JS functions
